requirements.txt
bcrypt==4.2.0
boto3==1.35.10
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
marshmallow==3.22.0
PyJWT==2.9.0
pytest==8.3.2
pytest-cov==5.0.0
pytest-env==1.1.3
pytest-xdist==3.6.1
python-dotenv==1.0.1
SQLAlchemy==2.0.32
